pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
# Check https://flit.readthedocs.io/en/latest/pyproject_toml.html for all available sections
name = "ansys-pre-commit-hooks"
version = "0.3.1"
description = "A Python wrapper to create Ansys-tailored pre-commit hooks"
readme = "README.rst"
requires-python = ">=3.9,<4"
license = {file = "LICENSE"}
authors = [
    {name = "ANSYS, Inc.", email = "pyansys.core@ansys.com"},
]
maintainers = [
    {name = "ANSYS, Inc.", email = "pyansys.core@ansys.com"},
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "importlib-metadata==7.1.0",
    "reuse==3.0.2",
    "GitPython==3.1.43",
    "Jinja2==3.1.4",
    "toml==0.10.2",
]

[project.optional-dependencies]
tests = [
    "pytest==8.2.2",
    "pytest-cov==5.0.0",
]
doc = [
    "ansys-sphinx-theme[autoapi]==0.16.5",
    "numpydoc==1.7.0",
    "sphinx==7.3.7",
    "sphinx-autodoc-typehints==2.1.1",
    "sphinx-copybutton==0.5.1",
]


[tool.flit.module]
name = "ansys.pre_commit_hooks"

[project.scripts]
add-license-headers = "ansys.pre_commit_hooks.add_license_headers:main"
tech-review = "ansys.pre_commit_hooks.tech_review:main"

[project.urls]
Source = "https://github.com/ansys/pre-commit-hooks"
Tracker = "https://github.com/ansys/pre-commit-hooks/issues"
Homepage = "https://github.com/ansys/pre-commit-hooks"
Documentation = "https://pre-commit-hooks.docs.ansys.com"


[tool.black]
line-length = 100

[tool.isort]
profile = "black"
force_sort_within_sections = true
line_length = 100
src_paths = ["doc", "src", "tests"]

[tool.coverage.run]
source = ["ansys.pre_commit_hooks"]

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
minversion = "7.1"
addopts = "--cov=ansys.pre_commit_hooks --cov-report term-missing -vv"
testpaths = [
    "tests",
]

[tool.towncrier]
package = "ansys.pre_commit_hooks"
directory = "doc/changelog.d"
filename = "CHANGELOG.md"
start_string = "<!-- towncrier release notes start -->\n"
underlines = ["", "", ""]
template = "doc/changelog.d/changelog_template.jinja"
title_format = "## [{version}](https://github.com/ansys/pre-commit-hooks/releases/tag/v{version}) - {project_date}"
issue_format = "[#{issue}](https://github.com/ansys/pre-commit-hooks/pull/{issue})"

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "dependencies"
name = "Dependencies"
showcontent = true

[[tool.towncrier.type]]
directory = "miscellaneous"
name = "Miscellaneous"
showcontent = true
